使用hutools工具替代
