+++ jexpose/src/main/java/com/hsiaosiyuan/jexpose/Colorize.java
