--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassA.java
@@ -1,5 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public class ClassA<TA> {
-  TA fieldA;
-}
