--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassB.java
@@ -1,11 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-import java.util.*;
-
-public class ClassB<TB extends ClassD, TB1 extends Date> extends ClassA<TB1> {
-  TB fieldB1;
-  List<TB> fieldB2;
-  HashMap<String, ArrayList<TB>> fieldB3;
-
-  private String pb1;
-}
