--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassD.java
@@ -1,5 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public class ClassD {
-  String fieldD;
-}
