--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassE.java
@@ -1,7 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-import java.util.Date;
-
-public class ClassE<T extends Date> extends ClassB<ClassD, T> {
-  String fieldE;
-}
