--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassF.java
@@ -1,6 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-import java.util.*;
-
-public class ClassF extends Date {
-}
