--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassFC.java
@@ -1,8 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public class ClassFC<TC1 extends ClassD, TC2> {
-  TC1 fieldTC1;
-  TC2 fieldTC2;
-  TC1 fieldB;
-  Object fieldA;
-}
