--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassG.java
@@ -1,4 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public class ClassG extends ClassD {
-}
