--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassH.java
@@ -1,4 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public class ClassH extends ClassI {
-}
