--- testjar/src/main/java/com/hsiaosiyuan/jexpose/ClassI.java
@@ -1,4 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public class ClassI extends ClassB {
-}
