--- testjar/src/main/java/com/hsiaosiyuan/jexpose/EnumC.java
@@ -1,14 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public enum EnumC {
-
-  C1((short) 1),
-  C2((short) 100);
-
-
-  private short v;
-
-  EnumC(short v) {
-    this.v = v;
-  }
-}
