--- testjar/src/main/java/com/hsiaosiyuan/jexpose/EnumE.java
@@ -1,15 +1,0 @@
-package com.hsiaosiyuan.jexpose;
-
-public enum EnumE {
-
-  E1("1", "12"),
-  E2("2", "22");
-
-  private String p1;
-  private String p2;
-
-  EnumE(String p1, String p2) {
-    this.p1 = p1;
-    this.p2 = p2;
-  }
-}
